requirements.txt
configobj
log
PyYAML
numpy
six
